refactor: scope lazy modal chunk failures with LazyModalSlot wrapper
